RTD template tweaks for css & toc layout
Top level page entries in toc were being pushed down a level if they
were under a category.  Move them back to l1 and return category headers
to where they used to be rather than demote top level page links under a
category to the same level as anchor links in a page that has not been
assigned to a category.